Fixed bug and simplify the code - Renamed CharArrayTextReader to Utf8BufferTextReader - Fixed length issue on netstandard version
